optimisation code

+++ app/controllers/pages_controller.rb
@@ -5,7 +5,15 @@ class PagesController < ApplicationController
   before_action :authenticate_user!
   def index
     @search_organismes = []
-    @organismes_last = @statut_user == '2B2O' ? Organisme.order(updated_at: :desc).limit(6) : @statut_user == 'Controleur' ? current_user.controleur_organismes.where(statut: 'valide').order(updated_at: :desc).limit(6) : current_user.bureau_organismes.where(statut: 'valide').order(updated_at: :desc).limit(6)
+    relation = case @statut_user
+               when '2B2O'
+                 Organisme.all
+               when 'Controleur'
+                 current_user.controleur_organismes.where(statut: 'valide')
+               when 'Bureau Sectoriel'
+                 current_user.bureau_organismes.where(statut: 'valide')
+                       end
+    @organismes_last = relation ? relation.order(updated_at: :desc).limit(6) : []
   end
 
   def mentions_legales; end
@@ -14,6 +22,8 @@ class PagesController < ApplicationController
   def plan; end
 
   def reset_all
+    redirect_to root_path and return unless @statut_user == '2B2O'
+
     OrganismeRattachement.destroy_all
     OrganismeMinistere.destroy_all
     OperateurProgramme.destroy_all
